Elbow-driven system for Raptor Reloaded
This work is a derivative of the following works, all but the last of
which are from the e-NABLE Community (www.enablingthefuture.org):
http://www.thingiverse.com/thing:596966
http://3dprint.nih.gov/discover/3dpx-000679
http://www.thingiverse.com/thing:476403
http://www.thingiverse.com/thing:261462
http://www.thingiverse.com/thing:261462

This adapter connects a standard Raptor Reloaded hand to a modified RIT
Arm cuff to allow transradial amputees with long forearms to use the
hand via an elbow-driven mechanism.

Due to an extremely busy lab schedule, we can currently only offer STL
files and photos of this device.

At present we do not have the bandwidth to provide any documentation or
instructions and will be unable to answer any questions.

We are providing these files for use by the e-NABLE community or anyone
else who would like to experiment with, build or improve upon the
design.

Thank you!